ENH: Reorganized module to allow for future growth
Reorganized the arch code to reside in univariate rather than in the root
Imported core routines in the arch module
Added temporary compatibility fixes with deprecation warnings
