+++ arch/__init__.py
@@ -2,7 +2,7 @@ import os
 
 from numpy.testing import Tester
 
-from .mean import arch_model
+from .univariate.mean import arch_model
 from ._version import __version__
 
 test = Tester().test
@@ -10,4 +10,3 @@ test = Tester().test
 def doc():
     import webbrowser
     webbrowser.open('http://arch.readthedocs.org/en/latest/')
-
