+++ arch/univariate/tests/test_volatility.py
@@ -5,14 +5,13 @@ import numpy as np
 from numpy.testing import assert_almost_equal, assert_equal, assert_allclose, \
     assert_array_equal, assert_raises
 
-
 try:
-    from arch import _recursions as rec
+    from arch.univariate import _recursions as rec
 except:
-    from arch import recursions_python as rec
-from arch.volatility import GARCH, ARCH, HARCH, ConstantVariance, EWMAVariance, \
+    from arch.univariate import recursions_python as rec
+from arch.univariate.volatility import GARCH, ARCH, HARCH, ConstantVariance, EWMAVariance, \
     RiskMetrics2006, EGARCH
-from arch.distribution import Normal, StudentsT
+from arch.univariate.distribution import Normal, StudentsT
 from arch.compat.python import range
 
 class TestVolatiltyProcesses(unittest.TestCase):
